Update patch set 2
Patch Set 2:

Thanks.

I'm aware it should be separated and was planning to.

I didn't notice that Adam had already pushed already as I'd assumed he was still working on something else still. Sorry for any confusion. I meant to set this as a "draft" but forgot to as it was for personal use for my build server :). Sorry for the trouble.

I'm about to abandon this.

Patch-set: 2