Update patch set 1
Patch Set 1: Code-Review-1

It's true that 2nd mic is enabled now, but the gain needs to be increased via "'DEC6 Volume':1:75" in order to record videos with similar loudness as on stock. This change unfortunately also enables 2nd mic on normal GSM/CSipSimple calls, where it causes loud noises after moving phone in hands. 2nd mic should be enabled only for recording videos and maybe recording music/voice, but not for calls.

Patch-set: 1