Update patch set 2
Patch Set 2:

Excellent, thanks for the feedback. I'm basing these changes on CAF files and Nexus 4 files, both of them have mono for voice recognition, in fact it's just a copy of "Handset" with a different ACDBID.

As for voice recording, my guess is that it probably shouldn't be using "Line" but it falls to that due to something else missing. Which app are you testing it with?

Patch-set: 2