Update patch set 2
Patch Set 2:

On stock voice recognition uses both mic simultaneusly, without that Google Now is unusable for lots of people, because the person has to be very close to the phone to activate the voice recognition feature.
Hmm, and Line is labeled as Speaker Tx - so maybe it's the right device, we need to check that.
I'm using this app: https://play.google.com/store/apps/details?id=com.andrwq.recorder.

Patch-set: 2