Update patch set 2
Patch Set 2:

Sorry for my previous comment, I was terribly wrong with that assumption. I was using stock Sony headphones, and these have inbuilt mic, so talking right to the mics on the phone obviously don't work.
By the way, I checked the audio-caf sources and it seems like "Line" device is actually the right device, so it's not broken or bugged.

Patch-set: 2