Update patch set 3
Patch Set 3: Verified+1 Code-Review-1

Changing to the CAF based ACDBIDs seems to make everything louder, ringtones/notifications are a bit distorted at full volume. Perhaps we can play with the volume values, in some places Sony set it to 100 where on CAF it is 66/68

I didn't add it in this patchset, but changing "Line" from DMIC1 to ADC1 (and adding volumes, just like handset) allows voice recording, but we need to find why DMIC1 isn't working in this case.

Patch-set: 3