Update patch set 1
Patch Set 1: Code-Review+1

I personally have a device with locked bootloader with CM running on top of stock kernel. I looked at the regular tree init.qcom.rc, and there it changed the pwrscale policy, which resulted in a reboot on .207 stock kernel. It is also the case in the normal CM - we can't change policy on a running system, only GPU governor can be changed. I previously thought we can do it on CM kernel, but nope, we can't.
The change fixes the bootloop that occured by using my commit that added CM perf profiles.

Patch-set: 1