Update patch set 9
Patch Set 9: Code-Review-1

I'd prefer not to rearrange everything at the same time as doing the bring up. Put the clean up (ie stuff that is simply moved around but doesn't change anything) in a separate commit so it's easier to review the actual changes.

Also remove the dependencies on the other commits so they can be merged.

Patch-set: 9