Update patch set 2
Patch Set 2: Code-Review+1

The Xperia SP Huashan has a LSM303DLHC LT Accelerometer that's used for orientation and software drivers driven by iNemo. Its compass is not used by the phone but by the iNemo, without the LSM compass the accelerometer is simply ignored (please check the Sensors.mk for this).

Then we have the AK8963 that is used as a dedicated Compass, using the SONY specific library. If you need proof of the SONY original sensors and the actual working ones on our builds, it can be provided.

Patch-set: 2