Update patch set 2
Patch Set 2:

The Xperia SP does have EAP support, for instance a lot of people I know on the french forums do use it over the FreeMobile networks. 

Stock ROMs supported it since the beginning and they complained about it on LP.

I made this change as a test and said I wanted to get their reports about it to merge.

I am still in the process of gathering their status, but if the feature was finally implemented in M and does actually work, let's give it back to the users who need it.

Patch-set: 2
Reviewer: Adrian DC <13589@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0