Update patch set 1
Patch Set 1:

Adrian, this is the wrong approach: It's fine to enable call recording in the device trees; to cover the legal issues we have by-MCC overrides in place in the dialer app (e.g. https://github.com/CyanogenMod/android_packages_apps_Dialer/blob/cm-12.1/res/values-mcc262/config.xml). So if the device is actually capable of doing call recording, this patch is fine.

Patch-set: 1