Update patch set 4
Abandoned

I'd like to get it working, but for the time being, none of the many changes & various test are reliable enough for me to merge it anytime soon.

Most implementations are working fine, and randomly trigger a stuck on boot with "ServiceManager: Waiting for service media.audio_policy...".

The first "complete" SHIM I worked on during November & committed in December has been used and bullet-proofed ever since, and never gave such an error. I'll get back on this if I can but it isn't trustworthy and properly debuggable for the moment.

Patch-set: 4
Status: abandoned