Update patch set 1
Patch Set 1:

thanks for the note. btw, looking at recoveryimg generation, i'm not entirely sure how mkelf works, but i'm guessing it is meant to replace mkbootimg for your device. if this is the case, then there's really no need for build-recoveryimage-target. you could remove this default routine and instead add this to the end of your manual routine:
$(hide) $(call assert-max-image-size,$(1),$(BOARD_RECOVERYIMAGE_PARTITION_SIZE))

something to note, build-recoveryimage-target outputs a sha1 hash of the first 20K (I think) of kernel, ramdisk, (and dt?) to recovery.id, which is then written to prop ro.expect.recovery_id. right now, this hash is likely incorrect for your device since it's based on the image that mkbootimg creates, not mkelf. so it might be in your interest to end use of mkbootimg (and hence the recovery.id file) since it stores invalid hashes for your recovery image anyways.

Patch-set: 1