Update patch set 3
Patch Set 3:

The LED waves now run fully in background,
the keycheck waits its 3 seconds during which you can press the keys.

If the key is pressed, the vibration state is now directly accessed in opposition to before where the wave had to be finished, and then it simply waits for the wave fork to be finished (or to be already done) to continue.

Hence Start, LEDs, keycheck pressed, vibrate, wait end of waves and continue. Exactly what it should be in all situations & timings.

Patch-set: 3