Update patch set 1
Patch Set 1: Verified-1

Kept only for future tests if the feature is improved by Google.

Current result makes no sense, the freeform windows are not shown on top of the desktop as one would expect, to watch a floating Youtube video for example.

Freeform windows switch the UI to an empty layout without desktop & one can only add new freeform windows with the bad-looking icon from multitasking.

The performances are not as pleasant & optimized as the Nougat multiwindow and definitely show no benifit on phones for now.

Patch-set: 1