Update patch set 2
Patch Set 2:

Yes I'll do it but in a future commit, per discussion with Alin Jerpelea, I'll look at doing it in a similar way for SODP as well at the same time.

I was also planning to split out the init rules to an INIT_RC like I did for my macaddrsetup or my lights_effects, but here again it involves everyone agreeing and could be blocking for those with incomplete sepolicies / init groups. We'll see if it's worth it.

Patch-set: 2